Closese #47 update mocks to match type definition (#51)
* Fixes #47 addSystem to world

This feature will allow a user to add a system to an already created
world via the `addSystem` function which consumes a system and adds it
to the list of systems that the world will execute on each tick.

* Closes #47 modify mocks so that tests pass

In my previous PR I had failed to update the mocks in __mocks__,
specifically for the world object. Since I updated its type definition
the tests were failing because I did not update the mock to match the
type definition.